Add defaults for SuspendedState instead of ignoring nil checks
